+++ src/network.py
@@ -3,6 +3,8 @@
 
 """Network related functions."""
 
+import collections
+import ipaddress
 import json
 import subprocess  # nosec
 
@@ -17,11 +19,37 @@ def get_router_id() -> str:
     return json.loads(out)[0]["prefsrc"]
 
 
-def get_network_interface() -> str:
-    """Get main network interface.
+def _get_network_interface(ip: ipaddress.IPv4Interface | ipaddress.IPv6Interface) -> str | None:
+    """Get network interface associated with the given IP.
+
+    Return:
+        Network interface name, or None if no route was found.
+    """
+    try:
+        out = subprocess.check_output(
+            ["ip", f"-{ip.version}", "-j", "route", "get", str(ip.ip)],  # nosec # noqa: S607
+            encoding="utf-8",
+        )
+        return json.loads(out)[0]["dev"]
+    except subprocess.CalledProcessError:
+        return None
+
+
+def get_network_interface(ips: list[ipaddress.IPv4Interface | ipaddress.IPv6Interface]) -> str:
+    """Get network interface associated with the given IPs.
 
     Return:
         Network interface name.
     """
-    out = subprocess.check_output(["ip", "-4", "-j", "route", "get", "1.2.3.4"], encoding="utf-8")  # nosec # noqa: S607
-    return json.loads(out)[0]["dev"]
+    names: collections.Counter[str] = collections.Counter()
+    interfaces: list[str] = [name for name in map(_get_network_interface, ips) if name]
+    names.update(interfaces)
+    name = (
+        names.most_common(1)[0][0]
+        if names
+        # fallback to the default route if none are found for VIPs
+        else _get_network_interface(ipaddress.IPv4Interface("1.2.3.4/32"))
+    )
+    if not name:
+        raise RuntimeError("unable to get network interface")
+    return name