+++ tests/integration/test_charm.py
@@ -109,6 +109,23 @@ def test_routing(juju: jubilant.Juju):
     juju.exec("ping 192.0.2.2 -I 198.51.100.2 -c 1", unit=test_b_unit)
 
 
+def test_keepalived_interface(juju: jubilant.Juju):
+    """
+    arrange: set the wireguard-a VIPs configuration.
+    act: read the keepalived configuration and determine the expected network interface.
+    assert: verify that keepalived used network interface.
+    """
+    juju.config("wireguard-a", {"vips": "203.0.113.2/24"})
+    juju.wait(jubilant.all_active)
+
+    status = juju.status()
+
+    for unit in status.get_units("wireguard-a"):
+        keepalived_config = juju.exec("cat /etc/keepalived/keepalived.conf", unit=unit).stdout
+
+        assert "eth0" in keepalived_config
+
+
 def test_bird_metrics(juju: jubilant.Juju):
     juju.deploy("opentelemetry-collector", channel="2/stable", base="ubuntu@24.04")
     juju.integrate("wireguard-a:cos-agent", "opentelemetry-collector")