+++ docs/changelog.md
@@ -7,3 +7,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
+
+## 2026-04-14
+
+### Changed
+
+- Select VRRP network interface based on VIPs configuration instead of using the default network interface.